Move store_info from FeatureListModel to DeploymentModel
