+++ featurebyte/models/feature_list.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 import functools
 from collections import defaultdict
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Set, cast
+from typing import Any, Dict, List, Optional, Set
 
 import pymongo
 from bson import ObjectId
@@ -24,7 +24,7 @@ from pydantic import (
 from typeguard import typechecked
 
 from featurebyte.common.validator import construct_sort_validator, version_validator
-from featurebyte.enum import DBVarType, SourceType
+from featurebyte.enum import DBVarType
 from featurebyte.models.base import (
     FeatureByteBaseModel,
     FeatureByteCatalogBaseDocumentModel,
@@ -34,17 +34,7 @@ from featurebyte.models.base import (
     VersionIdentifier,
 )
 from featurebyte.models.feature import FeatureModel
-from featurebyte.models.feature_list_store_info import (
-    BigQueryStoreInfo,
-    DataBricksStoreInfo,
-    DataBricksUnityStoreInfo,
-    SnowflakeStoreInfo,
-    SparkStoreInfo,
-    StoreInfo,
-    construct_store_info,
-)
 from featurebyte.models.feature_namespace import FeatureReadiness
-from featurebyte.models.feature_store import FeatureStoreModel
 from featurebyte.models.parent_serving import FeatureNodeRelationshipsInfo
 from featurebyte.models.utils import serialize_obj
 from featurebyte.query_graph.graph import QueryGraph
@@ -53,7 +43,6 @@ from featurebyte.query_graph.model.entity_relationship_info import (
     FeatureEntityLookupInfo,
 )
 from featurebyte.query_graph.node import Node
-from featurebyte.query_graph.node.schema import ColumnSpec
 from featurebyte.query_graph.pruning_util import get_combined_graph_and_nodes
 
 
@@ -369,7 +358,7 @@ class FeatureListModel(FeatureByteCatalogBaseDocumentModel):
     aggregation_ids: List[str] = Field(frozen=True, default_factory=list)
 
     # store info contains the warehouse specific info for the feature list
-    internal_store_info: Optional[Dict[str, Any]] = Field(alias="store_info", default=None)
+    feast_enabled: bool = Field(default=False)
 
     # pydantic validators
     _sort_ids_validator = field_validator(
@@ -583,57 +572,6 @@ class FeatureListModel(FeatureByteCatalogBaseDocumentModel):
         if feature_clusters_path:
             paths.append(Path(feature_clusters_path))
         return paths
-
-    @property
-    def store_info(self) -> StoreInfo:
-        """
-        Store info for a feature list
-
-        Returns
-        -------
-        StoreInfo
-        """
-        obj_dict = self.internal_store_info or {"type": "uninitialized"}
-        return cast(StoreInfo, construct_store_info(**obj_dict))
-
-    def initialize_store_info(
-        self,
-        features: List[FeatureModel],
-        feature_store: FeatureStoreModel,
-        feature_table_map: Dict[str, Any],
-        serving_entity_specs: Optional[List[ColumnSpec]],
-    ) -> None:
-        """
-        Initialize store info for a feature list
-
-        Parameters
-        ----------
-        features: List[FeatureModel]
-            List of features
-        feature_store: FeatureStoreModel
-            Feature store model
-        feature_table_map: Dict[str, Any]
-            List of offline feature table map from source table name to actual feature table
-            (could be a source table or a precomputed lookup table)
-        serving_entity_specs: Optional[List[ColumnSpec]]
-            List of serving entity specs
-        """
-        store_type_to_store_info_class = {
-            SourceType.SNOWFLAKE: SnowflakeStoreInfo,
-            SourceType.DATABRICKS: DataBricksStoreInfo,
-            SourceType.DATABRICKS_UNITY: DataBricksUnityStoreInfo,
-            SourceType.SPARK: SparkStoreInfo,
-            SourceType.BIGQUERY: BigQueryStoreInfo,
-        }
-        if feature_store.type in store_type_to_store_info_class:
-            store_info_class = store_type_to_store_info_class[feature_store.type]
-            store_info = store_info_class.create(
-                features=features,
-                feature_store=feature_store,
-                feature_table_map=feature_table_map,
-                serving_entity_specs=serving_entity_specs,
-            )
-            self.internal_store_info = store_info.model_dump(by_alias=True)
 
     class Settings(FeatureByteCatalogBaseDocumentModel.Settings):
         """
