+++ featurebyte/models/feature_list_store_info.py
@@ -40,7 +40,6 @@ class BaseStoreInfo(FeatureByteBaseModel):
     """
 
     type: StoreInfoType
-    feast_enabled: bool = Field(default=False)
 
     @classmethod
     def create(
