+++ tests/unit/models/test_feature_list.py
@@ -37,12 +37,12 @@ def feature_list_model_dict_fixture():
         "catalog_id": DEFAULT_CATALOG_ID,
         "relationships_info": None,
         "features_entity_lookup_info": None,
-        "store_info": None,
         "supported_serving_entity_ids": [],
         "block_modification_by": [],
         "description": None,
         "is_deleted": False,
         "aggregation_ids": [],
+        "feast_enabled": False,
     }
 
 
@@ -105,7 +105,7 @@ def test_feature_list_model(feature_list_model_dict):
     assert loaded_old_feature_list == updated_feature_list
 
     # check that feature list store info for older record
-    assert loaded_old_feature_list.store_info.feast_enabled is False
+    assert loaded_old_feature_list.feast_enabled is False
 
 
 def test_feature_list_namespace_model(feature_list_namespace_model_dict):
